+++ NEWS.md
@@ -9,5 +9,5 @@
  * Added code for plotting fitted data
  * Added predict() function.
  * Added function for filtering data based on the leaves in a tree and for mapping between samples and populations (leaves).
- 
+ * New optimisation code based on solving linear systems for those variables that are linear.